init: use viper for config
