+++ systemgo.yaml
@@ -1,0 +1,1 @@
+Debug: true
